#24: [fix] 画像アップロードのロジックをhooksディレクトリにわけ、ロジックと描写が分かれるようにした
